[cli] remove deprecated `certs renew` command (#12248)
This was deprecated 6 years ago and has not done anything beyond warn
for quite some time. `major` feels too heavy for this change although it
does _technically_ alter the public interface. But I find it hard to
believe anyone was relying on this specific error message after such a
long period, so went with `minor`