[cli] mock server should fail harder when a route is not implemented (#12311)
Our test mock server returns a `404` status when we've failed to implement a route necessary for the test to run. I think this is incorrect and should in fact be a `500` (and possibly even throw immediately?): the CLI code interprets 404s to mean a resource is missing and doesn't distinguish between this and "you made a request to a completely unknown location".

Thankfully we only have ~one~ two tests relying on this behavior.