[cli] add telemetry tracking for vercel dev invocations (#12349)
This one was a _beast_ to get right because of how much
filesystem/environment mocking needed to occur, how little already
existed in the repo, and how little I knew about doing these in vitest.
Some notes!

* added `memfs` which is the library vitest
[recommends](https://vitest.dev/guide/mocking.html#file-system) for
mocking the file system. Their example involves mocking for every test
which I'm not sure yet we'd want do. Figuring out the right incantation
to fully mock the filesystem in just one test file was an adventure.
* Updated `vitest` to latest so we can [stub environment variables to
`undefined`](https://github.com/vitest-dev/vitest/pull/6359). Our
version of vitest has `unstubAllEnvs` but that seemed like it would just
paint us into a weird corner. We use `process.env.__VERCEL_DEV_RUNNING`
as a lock to prevent someone from running `vercel dev` twice in one
process but in tests this is exactly what occurs. Outside of tests we
use rely on the process exiting to do the cleanup for us.
* I mocked the entire `DevServer` class _very_ basically. We don't want
to actually run it but when someone comes back to add more unit tests
they should add in more spys to ensure the server is being invoked with
the arguments we expect and methods are called in the correct order.

We do have our own filesystem mocking [in the
repo](https://github.com/vercel/vercel/blob/d392591ebdb28b809a9411a6930d6ede397df202/packages/fs-detectors/test/virtual-file-system.ts),
but not as part of `cli`. Extracting that seemed a bit too heavy and IMO
we should use a third party library for this anyway.

---------

Co-authored-by: Jeff See <jeffsee.55@gmail.com>
Co-authored-by: Austin Merrick <onsclom@onsclom.net>
Co-authored-by: Nathan Rajlich <n@n8.io>