[cli] Add telemetry for `vercel init` (#12371)
This one is a slightly odd duck because the "example" positional argument will track the literal value, but only for valid template names. If the user typos or supplies an otherwise unknown value, we will redact the data from the telemetry. Since the list of known template names requires an API call, this tracking happens further into the command's logic than the others.