extract vc pull logic from entrypoint (#12378)
This is a very basic extraction of the core `vc pull` logic out of the command invocation for use in the build command invocation. This gives us a place to add telemetry later in only the `vc pull` invocation logic so that invoking `vc build` does not trigger `vc pull` telemetry metrics.

I manually tested this to make sure I didn't break it.