ECO-444: Add remove subcommand for integrations command (#12291)
## Summary

This PR adds a new sub-command to integration: `remove` (also available as `rm`) which allows a user to uninstall a marketplace integration from the current team. This PR also adds a new command `vc integration-resource` with sub-commands `disconnect` and `remove` (also available as `rm`), which allows a user to unlink a project from a (marketplace integration) resource and to delete a (marketplace integration) resource.

## Background

This change is related to #12233 and #12251. As we continue to try and improve the developer experience of interacting with the marketplace via the CLI, it's important that we offer the basic general CRUD activities seen in CLI tools of `add`, `list`, and `remove` (as well as `open` for how users perform update actions on the resources day-to-day). This PR adds the `remove` functionality to finish off the set of the basic CRUD operations.

Original designs for the CRUD commands lacked detail on a `remove` command. This PR introduces the idea of removing an integration from a team, removing a resource from a team, and removing resources from projects. Since this is a lot of different destructive actions, we thought it best to break apart into multiple commands instead of overloading a single `remove` command (as this PR originally proposed). This way the commands have clear, singular uses, and it's more difficult to accidentally make unintended destructive changes.

## How to test

Build and run locally (or use `pnpm dev` instead of `vc` when in the `packages/cli` folder). You'll also need marketplace integrations installed on a team, and resources for said integrations.

```
vc integration remove <integration>
--
vc integration-resource remove <resource> [--disconnect-all]
--
vc integration-resource disconnect <resource>
vc integration-resource disconnect <resource> <project>
vc integration-resource disconnect <resource> --all
```