[cli] redacting regions until we determine method to get list (#12553)
We would like to track regions since it's a Vercel known list, but we
don't have a way to easily get that list at the time of the tracking
method call here. For now we'll redact this until we have a way to get a
list we can scan through.