[python] fix ASGI response streaming (#12610)
`asyncio.Queue` requires a `loop` parameter on Python < 3.10. Copy the
same logic as for non-streaming using the `_use_legacy_asyncio`
variable.

This was only discoverable when an ASGI response would stream something,
which none of the test were doing. I've add a new probe for the
`28-pydantic` test to stream a response.