[cli] Swap out custom CI code for `ci-info` package (#12569)
We're currently using some hand-crafted code with a data object
copy/pasted from `turbo` in out CI platform detection for telemetry.
Someone suggested we use the existing `ci-info` package but there was no
way to a get stable identifier for each platform.

You _could_ do:

* `ci.name` but it was specifically noted this could change
* `ci.<SOME_CONST>` but that involved us mapping to ids ourself, which
we feared would get out of date

https://github.com/watson/ci-info/commit/81fd993 adds a `ci.id`.

Testing this is a bit wonky because the `ci.id` is set on package import
once, so the mocking was complicated. `vi` hoists it's package mocking
to the top of a file using static analysis, so can't be defined twice.
The workarounds for this all apply to mocked _functions_ but not
imported constants.

The only thing I could think of doing was moving each of these cases
into their on file 🤷