[cli] Gracefully report error in `vercel link --repo` when a new Project fails to be created (#12616)
Fixes a bug in the `vc link --repo` command where an API error would be
ignored, causing the command to fail further down in the logic. Now the
API error is properly propagated and logged, and the command exits with
a proper exit code.

Includes a unit test of the of this failure scenario, and an additional
bonus unit test for the success case of `--repo` creating a new Project,
since that was missing before.