[cli] add eslint prefer-const rule and fixes (#12654)
Adds the eslint `prefer-const` rule to the `cli` package. Fixes were
applied, with the exception of `let [item1, item2] = <var>` pattern
matching where one of the items _is_ being reassigned and therefore
can't be a const.

In the future, it would be worth investigating these uses to see if a
non-reassignment version of the code can be refactored.