[cli] remove name from CustomEnvironment type (#12709)
We'd like to remove `CustomEnvironment`'s `name` property at the API
before going GA. Although technically a breaking change, we're just
going to communicate to beta/preview customers about this.

This removes its use in the CLI. We only ever submitted `id` / `slug` to
the API and only used `name` for more humane labeling. I've reproduced
this with `title(env.name)`, but am open to just never showing this
field.