[tests] remove project settings mechanism for dev tests (#12724)
The `projectSettings` option for `testFixtureStdio` is used only once.
This was likely more useful at some earlier point or was required
because we didn't previously have stable projects we tested against
(each test run created accounts/projects on its own).

We never do _cleanup_ for this either, so the entire project already has
this value set and has for 168 days.

I'd be amenable to leaving in place since it does colocate the specific
setting under test with the test itself which I like. But practically
this isn't doing anything other than adding a noop HTTP call to a single
test.