Enable new storage integrations via wizard (#12733)
We moved from using a `type` field to a `protocols` object with some
additional information for determining the category of an integration.
New storage products don't have the old `type` field, instead using the
new protocols, but are otherwise still storage products with the same
wizard setup needs. New integrations that would otherwise be fine to use
the wizard for setup were directing users to the dashboard to complete
the installation process (undesired behaviour).

This PR adds in the typing for the protocols currently in use (as of
this PR, just storage), so that our newest integration products can be
added purely via the CLI.