[remix] Remove `@remix-run/dev` dev dependency
It was being used for two purposes:

 1. Importing types from an internal sub-path
 2. Determining the maximum safe version of the forked compiler to use

Point 1 is solved by copying the relevent types directly into the Remix
builder package

Point 2 is solved by moving that max version out of `devDependencies`
and into a top-level `package.json` property. The reason for the
placement (i.e. instead of hard-coding the version into the source code)
is so that the value can be easily updated via the update cron job
(if we wanted to go the source code route, we could make the cron job
update the source code via AST transformation at some point in the
future, but this works for now).