[next] Root Params (#12764)
Added support for root params on pages powered by partial prerendering.
This enables shell generation for pages with the root params. Now, when
partial prerendering is enabled for a route with a root parameter, such
as structure like this:

```
/[lang]/layout.tsx
/[lang]/page.tsx
```

The `lang` key will be included in the generated shell that's returned.
For parameters that were prerendered (via `generateStaticParams`), they
will already have more complete shells generated. This specific blocking
route is only used when routes are served that have unknown root params.