Refactor build-utils usage to reuse lockfile path from scanParentDirs (#12749)
We detected the lockfile file name in `detectPackageManager()` using
just `cliType` and `lockfileVersion`. While this was fine for the
previous package manager detection, we can't know whether the lockfile
name is `bun.lock` or `bun.lockb` with just that info. This resulted in
some [hacky
logic](https://github.com/vercel/vercel/pull/12740/files#r1882591335) to
support `bun.lock`, and this PR fixes that.

This refactors `detectPackageManagers()` to take in the already detected
lockfile name (detected from `scanParentDirs` in our usage), which means
we also need to argument drill through
`getPathOverrideForPackageManager()` and `getPathForPackageManager()`,
update all the builders using these functions, and update lots of tests.