[build-utils] document `NODE_VERSIONS` implicit sort in comment (#12865)
`NODE_VERSIONS` is assumed in other parts of `build-utils` to be sorted
from newest Node version to oldest. This PR adds a comment to document
that assumption.

For example,
https://github.com/vercel/vercel/blob/a567aaddc90016975583f442ecb61a4f1d0eefcb/packages/build-utils/src/fs/node-version.ts#L73-L89
assumes that the latest version _is_ the first version. If the sort was
disrupted in `NODE_VERSIONS` this function would start providing wrong
behavior.

We could make this assumption explicit by sorting `NODE_VERSIONS` before
checking for latest, but this adds a performance penalty to the Node
version lookup. Without tracing this functions use, I can't tell if this
would be a performance regression on the critical path.