[cli] update warning in ls --update-required for clarity (#12864)
Closes #12856.

The existing logic assumes that if a `discontinueDate` has been set for
a given version of node, that version of node must be deprecated. It was
also easy to mistake the previous code as displaying the warning only
_after_ the discontinue date, which it never did.

This PR:

- Clarifies the language communicating the deprecation to users. It's
not that the node version _will be_ deprecated. The node version is
_already_ deprecated and is at risk of being discontinued.
- Adds an identifier to the logic checking for a deprecated node
version, making it clear that the date check is looking into the future.
- Adds a comment clarifying the logic further, and documents the
assumption that a `discontinueDate` means that a node version is
deprecated.
- Adds a further comment noting that if this assumption is invalid,
additional logic may be required. Specifically, in the case where we
have deprecated a node version, but don't yet know its discontinue date.