[build-utils] add state property to NodeVersion  (#12883)
# Overview

This PR adds a `state` property to the `NodeVersion` type. One of
`'active' | 'deprecated' | 'discontinued'`. This changes `NodeVersion`
from an interface with optional fields to a tagged union with `state` as
a discriminator.

Now `discontinueDate` is a required field only on
`DeprecatedNodeVersion` and not a field on the other two alternatives.

## Goal

The `state` property makes it clear at a glance what state a given
NodeVersion in `NODE_VERSIONS` is in.

Previously, code assumed that if a `discontinueDate` had been set, a
version was deprecated, but this assumption was not explicit in the
code.

This PR makes that assumption explicit.

## Work Done

- Updated the `NodeVersion` type to be a type union of tagged
alternatives, one of each state.
- Updated `NODE_VERSIONS` to include the `state` property.
- Updated the helper functions `node-version.ts` to account for the
`state` property.
- Updated some language in the warnings printed by `node-version.ts` to
use our newly considered terminology around deprecation and
discontinuation.
- Updated project list `--updated-required` to use the `state` property.
- Updated `RubyVersion` which depended directly on `NodeVersion` being
an extensible interface instead of a type union.

## Questions

- Do we want another field name besides `state`? If so, what?
- Do we want `RubyVersion` to still be dependent on `NodeVersion`?
- Is there any reason to include `discontinueDate` for `discontinue`
state Node Versions?