[examples] Update Blitz to 2.0 (#12911)
Blitz 0.x won't build on Node > 16 so we need to bump this. Can't quite
tell why but every Blitz 2.x generator and every option in that
generator fails to build locally. We initially thought this was only
[related to latest
2.2.0](https://github.com/blitz-js/blitz/issues/4408), but it's all of
them.

This one was hand massaged from a 2.0 generated project.

Interesting installing dependencies during a deploy via `pnpm` fails:
![CleanShot 2025-01-24 at 15 35
10@2x](https://github.com/user-attachments/assets/03552f18-6e88-4626-b8cd-eb5ef6fbc86c)

So this has a npm lockfile instead.