[remix] Add support for React Router v7 (#12904)
Updates the `@vercel/remix-builder` Builder package to support React
Router v7 apps. Implementation-wise, RRv7 is handled almost identically
as Remix v2 + Vite, but with the name "remix" changed to "react-router"
in a few places.

All the same features that are supported on Vercel when using Remix are
now also supported when using React Router (when combined with the
Preset, see #12903).

**NOTE:** E2E tests will be added in a follow-up PR after the new
framework preset is added, so that autodetection will work in the tests.