[next]: ensure defaultLocale redirect doesn't conflict with user redirects (#12916)
When `localeDetection` is turned off, usually the user intends to handle
redirects themselves, or doesn't want them at all. However because of
the rewrites that we insert for `defaultLocale` handling, it can clobber
the user-provided redirects. eg the rewrite to `/${defaultLocale}` would
then flow through to the user's redirect handling to result in
`/${defaultLocale}/${redirectLocale}` which wouldn't have been the
intention.

When `localeDetection` is turned off, we shouldn't only attempt the
defaultLocale rewrite handling when we've already processed all other
routes (ie in the `handle: miss` case)

Closes NEXT-3975

---------

Co-authored-by: JJ Kasper <jj@jjsweb.site>