[node][remix][routing-utils] log diff between current and updated versions of `path-to-regexp` (#12926)
Using [pnpm aliases](https://pnpm.io/aliases), this PR installs two
different versions of `path-to-regexp`. Both are run and the resulting
regex is compared. If they are different, an innocuous log message is
sent to the build logs.

This allows us to test out the new version and be notified of any
breaking changes it would have for our users.

---

Testing: This will be manually tested in the build container before
release.

---------

Co-authored-by: Erika Rowland <130414394+erikareads@users.noreply.github.com>