[tests] Remove hard-coded node version from `packages/build-utils` tests (#12940)
We decided to remove any "engines" value in fixtures that are not
directly related to the test in question. Node 16 is being removed so
starting by removing these since they'll stop working on Friday.

There _are_ two fixtures that I believe are intended to test node 16
specifically:

*
[`20-npm-7`](https://github.com/vercel/vercel/blob/7dd88ef2b6dabf231056a181de4604d042c46c49/packages/build-utils/test/fixtures/20-npm-7/package.json)
*
[`15-npm-8-legacy-peer-deps`](https://github.com/vercel/vercel/blob/7dd88ef2b6dabf231056a181de4604d042c46c49/packages/build-utils/test/fixtures/15-npm-8-legacy-peer-deps/package.json)

so these will be removed in a different PR.

Co-authored-by: Sean Massa <EndangeredMassa@gmail.com>