[next]: ensure locale prefixed dynamic routes come before non-prefixed (#13003)
With the experimental flag added in #12916 we added handling to add
non-localized redirects for dynamic routes, because we moved the
defaultLocale handling to the `handle: miss` case. However, we need to
add these redirects after the less-specific localized routes so that
those ones are matched first before the fallback case.

---------

Co-authored-by: JJ Kasper <jj@jjsweb.site>