[tests] Remove local tests of corepack behavior (#13014)
NPM changed its singing keys, which corepack uses to verify that the
downloaded package manager is legit. This means that corepack has be
updated _on our platform_ to install newer packages that were signed
with the newer key.

In effect: You could not install `pnpm@10.1.0` or `pnpm@9.15.4` with
`corepack@0.30.0` or below briefly on platform before we fixed.

This is _still_ true for tests running in CI and customers using this
locally.

These unit tests covered the general corepack selection behavior when we
assumed the corepack that shipped with node would exactly match what we
used on platform, which is no longer true.

This behavior now needs to be tested with actual deploys as part of the
platform build process, not the CLI itself.