[build-utils] [tidy] extract checkIfAlreadyInstalled helper to clarify intent (#13060)
Prior to this PR, we had logic in `runNpmInstall` that prevented
duplicate installs based on the same `packageJsonPath`. The code was a
nest of conditionals that made it difficult to follow what it was trying
to accomplish.

This PR extracts the conditional logic into a couple helper functions,
and refactors `runNpmInstall` to make it clear what's happening: we're
returning early if a default install has already happened for a given
packageJsonPath.

---------

Co-authored-by: Austin Merrick <onsclom@onsclom.net>