[remix] Fix build failure when using an index route with a pathless layout (#13108)
Handle the case when an index route is used within a layout route. For
example, with React Router:

```ts
import {
  type RouteConfig,
  index,
  layout,
  prefix,
} from '@react-router/dev/routes';

export default [
  layout('./layouts/app-layout.tsx', [
    ...prefix('/', [
      index('./routes/home.tsx'),
    ]),
  ]),
] satisfies RouteConfig;
```

Previously this would be failing at build time with the error:

> #### Build Failed
>
> An unexpected error happened when running this build. We have been
notified of the problem. If you have any questions, please contact
Vercel Support https://vercel.com/help

The failure was caused by a function with an empty path attempting to be
created, which is not handled on the build side and causes the
deployment to fail.