[next]: Read expire values from prerender manifest (#13087)
In https://github.com/vercel/next.js/pull/76207 we've added
`initialExpireSeconds` (and `fallbackExpire`) to the prerender manifest,
analogous to `initialRevalidateSeconds` (or `fallbackRevalidate`). Both
the revalidate and expire values together allow stale-while-revalidate
semantics to be applied to static routes and fallback routes for ISR.

The configured expire values are added as `staleExpiration` property to
the prerender outputs. The naming is a bit unfortunate, but that's a
result from the revalidate value being transferred to the `expiration`
property.

- [x] The added integration tests depend on a new Next.js canary version
being released after https://github.com/vercel/next.js/pull/76207 is
merged.