Allow all integration categories for resource provision, send to dashboard on non-subscription plan selection (#13220)
Previously we were filtering which integrations could be installed via
the CLI based on their product type. This is problematic as we scale, as
it requires a CLI change & deployment every time a new type rolls out.

Originally, we set it up this way to prevent breaking the CLI as new
requirements were discovered in different integration product types.
What we've found instead is that the generic integration concept is fine
across products, but the billing plan structure & system may change. So
instead of having to make changes every time a new product type comes
out that should be supported by default, we're instead switching to
sending users to the dashboard for unsupported billing plan types.

Today, the `prepayment` billing plan type requires additional work to
support provisioning via the CLI, as it has a different flow from
subscription-based plans. It is also the only other plan type beyond
`subscription`, but we'll filter all non-subscription plans to future
proof against potential future plan types.