[cli] allow configuring `architecture` in `vercel.json` (#13344)
The `architecture` of a Vercel Function (`x86_64` or `arm64`) has been
configurable for ages in the Build Output API:
https://vercel.com/docs/build-output-api/primitives#serverless-function-configuration

However, this option was not exposed inside the `vercel.json`
`functions` configuration option:
https://vercel.com/docs/project-configuration#functions

Similarly to e.g. `runtime` or `memory` in `vercel.json` `functions`
config, allow configuring the architecture per function to override the
default of x86_64. For example:

```json
{
  "functions": {
    "api/test.js": {
      "architecture: "arm64"
    }
  }
}
```

---------

Co-authored-by: Dimitri Mitropoulos <dimitrimitropoulos@gmail.com>