[cli] add try/catch around login prompt attempts (#13399)
A bit of an odd bug. We handle a number of errors inside `doLoginPrompt`
and will exit with a code returned from there, but in at least one case
(`Ctrl-C` to exit the prompt) we do not detect and will exit with `0` in
this case.

This should address the issue.