feat: introduce `getVercelOidcTokenSync` (#13429)
`getVercelOidcToken` doesn't need to be async, and there are places we'd
like to use it where it would be more convenient if it were sync.
Therefore, this PR introduces a new, synchronous version:
`getVercelOidcTokenSync`.

There are still some places that could benefit from an asynchronous
`getVercelOidcToken`, and so this version is still there, and not marked
as deprecated.