[vercel cli] Prefer fetching microfrontends config from project over a deployment (#13485)
This PR modifies the logic for the `vc microfrontends pull` to first
prefer pulling the MFE config from the most recent production deployment
given the project ID if a `--dpl` parameter is not specified. This fixes
a case where it would be an error for a new project to be added to a
microfrontends group before a new deployment is generated for that
project.