[frameworks] use correct output dir for nitro (#13541)
Unsure _why_ but starting in `44.2.9` Nitro deploys stopped working
correctly. I believe this is the root cause of the error:

```
***CLI Stack Trace:***
Error: The "path" argument must be of type string. Received function Promise
at Now.handleDeploymentError 
```

The current configuration returns the `Promise` constructor when it
should return a string value that is a path.

What in `44.2.9` is beyond my ken, but I believe this addresses it. If
not, customers can set `'public'` as their Output Directory value in
their project's `settings/build-and-deployment` page. I verified that
works correctly.

---------

Co-authored-by: Chris Olszewski <chrisdolszewski@gmail.com>