[cli] Use streams to upload Blob files instead of reading it all in memory (#13523)
Before this commit, we would read the FULL file before sending it to
Vercel
Blob. This can make your memory go wild.

Now we will stream the file from disk to Vercel Blob directly.
I also updated the unit tests so they do not rely on mocking the
filesystem calls which are an internal implementation, we only need to
check input => output. Ideally I could check the output read stream and
verify its content but that's not so easy to do. The ultimate play here
is to avoid unit tests for these simple cli commands and only have
integrations tests, the real ones.