fix(react-router): avoid injecting 404 route if additional runtimes are found (#13571)
Currently if additional functions are declared with other runtimes e.g.
Python we inject a `/api(/.*)? -> 404` route to prevent `/api` from
returning the directory (https://github.com/vercel/vercel/pull/5064).
This is unnecessary for React Router which handles this for us and
overfits resulting in declared React Router routes returning a 404.