[python] support `/_vercel/ping` endpoint for Fluid (#13589)
Add support for the `/_vercel/ping` endpoint to replace the
`handler-started` IPC message introduced in
https://github.com/vercel/vercel/pull/12850.

We first need to land this change (no-op), then allow our runtime to use
this new endpoint instead of the `handler-started` message, then clean
up the previous logic. We cannot test this change in CI yet (until we
land the runtime change), but I've tested it manually.