Add hono to framework-detection and associated builder (#13594)
This adds framework detection and an associated builder for Hono so that
it can be used in the root as opposed to needing to be in the `api`
folder. The framework detection is pretty strict at the moment to ensure
we don't get any false positives for non-hono apps.

- Checks that hono is a dependency
- Checks that there is a file (either `index.{js,ts}` or
`src/index.{js,ts}` that has some sort of hono import/require

The builder is a light wrapper around the node builder that uses a shim
file to tweak the handler functions. This allows users to use the normal
`default export app` from the Hono [docs](https://hono.dev/docs/)

This also doesn't affect code inside the `api` folder. If you have a
hono app there, you'll still need to user the `hono/vercel` handler. TBD
on what we should do about that.

More tests to follow, but much of our test infra is based on the builder
being available on the build container, so chicken & egg situation