[next] only route prefetches to static files if header is '1' (#13547)
x-ref: https://github.com/vercel/next.js/pull/81088

"Runtime prefetches" (WIP in the linked PR) set `next-router-prefetch:
2` instead of the usual `next-router-prefetch: 1`, and should always hit
a lambda, because we render the result dynamically. This mostly works
fine with what we generate here, except for a couple rules that only
check for the presence of the header, not its value. This PR adjusts
them.

Existing versions of Next never set the header to anything other than
`1`, so this is backwards compatible.

Co-authored-by: JJ Kasper <jj@jjsweb.site>