Support fetchable apps out of the box for Node dev server (#13664)
This removes the shim file now that BOA entries with a `.fetch` property
are supported. There were some issues with the shim when the compiler
compiled to CJS which will be resolved by this.

It also fails the built when TSC finds an error. Right now, this
behavior is only turned on for Hono apps so as to not disrupt existing
behavior.