fix(go): Respect patch version 0 and update minor version entries (#13672)
This PR fixes a bug in our parsing of exact go versions in `go.mod`. If
the user specified a version with a zero patch e.g. `1.24.0` we would
not respect it and instead lookup a patch in the minor version map.

This PR also bumps our minor version map to their latest patches.