Add express builder (#13711)
This is essentially an identical copy of the hono builder. I plan to
abstract this into a shared package as we get validation about the
approach here. There's also the possibility that neither package will be
necessary, as the node builder can likely handle the bespoke entrypoint
logic that these "builders" do.

There's also a follow up to add this to the cli package, which I'm not
doing in this PR because when I did it with Hono, but package failed to
release properly and didn't cause the CLI to fail releasing.