Add support for custom build scripts for hono (#13724)
Adds support for a custom build script for Hono projects, but the Hono
builder still needs to run on the entrypoint (src/index.ts, server.ts,
etc.) so if the build command compiles source code into a dist
directory, we currently wouldn't detect that.

However for turborepo apps, a `build` script may build workspace
dependencies, in which case this is needed.