[cli] add supportsCancellation option to functions (#13807)
Cancellation of requests to Fluid / N1 functions running on Lambda
requires breaking changes to how the function lifecycle works. For this
reason, and to allow rapid but limited rollout of the feature,
configuration of which functions be cancelled will exist in
`vercel.json` in the `functions` blocks.

This PR adds support for the optional boolean `supportsCancellation`
field:

```json
{
  "functions": {
    "api/cancellable.js": {
      "supportsCancellation": true
    }
  }
}
```

This is then written to the `.vc-config.json` file when building.

```json
{
  "handler": "api/cancellable.js",
  "runtime": "nodejs20.x",
  "architecture": "arm64",
  "environment": {},
  "launcherType": "Nodejs",
  "shouldAddHelpers": true,
  "shouldAddSourcemapSupport": true,
  "awsLambdaHandler": "",
  "supportsCancellation": true
}
```