[next] add supports cancellation (#13841)
# Problem

The Next.js builder does not support the `supportsCancellation` option
when bundling.

# Solution 

Ensure that the `supportsCancellation` option is read and passed through
to the `NodejsLambda` constructor.

*Question*

Technically the `supportsCancellation` flag does not require splitting
the bundle, as whether to allow cancellation or not is a routing-time
decision rather than a deployment decision. Putting the option in the
deployment path metadata just tells proxy/s-f-r that cancellation is
allowed.

Is there a way to support this in the `.vc-config.json` file of
supported routes without also using it as a differentiator in bundling?

---------

Co-authored-by: Craig Andrews <craig.andrews@vercel.com>
Co-authored-by: Craig Andrews <44933829+craigandrews@users.noreply.github.com>