Support package.json#main field as the entrypoint (#13846)
The changes the way we detect the entrypoint for Express and Hono apps.
Allowing the value to be provided as `package.json#main`. While working
on this, I realized a few things that led to a slightly large diff:

- Express supports a `views` folder for rendering engine templates. This
value is configurable, but don't have a great way of knowing that at the
moment, so for new `views` is included by default for express apps.
- It's better to just use `glob` than to worry about windows / posix
filepath issues, so that was changed around to be more simple
- We previously supported looking for an entrypoint in an output
directory if the project set one, however you still needed a valid
entrypoint to exist before we ran your build script (if provided). So
this PR changes the order in which we find your entrypoint, always doing
it after any install/build scripts have finished.