chore(cli): telemetry on login state (#13886)
This re-lands #13855 with a different approach. We had to revert
(#13872) the original approach as it caused the login process to hang
after a successful login for some reason, which I suspect was due to
some changes in the interruption handling.

This PR passes the `telemetry` client to the `future` function instead.
Note that there is a refactor coming in #13802 and this approach plays
nicer with that anyway.