fix(cli): use introspection endpoint during token refresh (#13888)
During an authenticated request, we check if the access token is still
valid, and automatically refresh it if not.

There was an issue on this line:


https://github.com/vercel/vercel/blob/fee07858ef049cc3ae71d3a9b66b42139f06c8a5/packages/cli/src/util/client.ts#L79

We assumed the `expiresAt` value is in milliseconds, but in fact it is
in seconds:


https://github.com/vercel/vercel/blob/fee07858ef049cc3ae71d3a9b66b42139f06c8a5/internals/types/index.d.ts#L45-L49

This led to the access token being refreshed all the time, even if it
was still non-expired.

We can get rid of this field by using the introspection endpoint
instead, which considers more than just the token expiry anyway.