[cli] Fix vc deploy to show full build logs on build error (#13901)
This fixes a bug where sometimes the `vc deploy` with build errors do
not return the full logs:
https://linear.app/vercel/issue/FLOW-3934/vercel-deploy-builds-logs-do-not-contain-full-build-logs

We receive the build error before the logs are ready to grab. Ideally,
we'd want some general way to know when logs are done, but in this case
we know what the last log line will be, so we can watch the logs until
we see that.