[next-builder] Don't explicitly exit(1) the process on error bubbling to the root of the handler
Instead let things bubble to the proper global error handling facilities (which e.g. in the case of Fluid keep the process alive until other requests finished).

In practice, it is unexpected that errors bubble up to this stage, but it can happen in true exceptions or when next.js has a bug (which is how we found this).