feat(oidc): add conditional export for browsers (#14005)
This is just a base of discussion. It would make it easier for us to use
`@vercel/oidc` in the AI SDK. And if we go down this path, we could also
make conditional exports for Bun, Deno, and other run times in case they
need different code for file system access, OS-specific APIs or
environment variables.

If we want to go ahead with this PR, I'd like to merge
https://github.com/vercel/vercel/pull/14004 first, then update this PR
accordingly, so that the default and browser entry points don't diverge.

See also https://github.com/vercel/ai/pull/8975