Recover from release error (#14027)
### Background

The cause for the release failure was that changesets for examples are
ignored by config, which broke the release flow.

Why and how that resulted in a release failure is laid out perfectly in
@tomdale's comment:
https://github.com/vercel/vercel/issues/14023#issuecomment-3359156378

### Summery of changes

This pull request restores a valid changeset state prior to the merge of
the last "Version Packages" pull request.

476567e168e0f4a49059e1f33f75c6cf4342d18a reverts all changes from #13989
2ee9eb5520b330fcf0a90169a559f173f6d82f19 removes the two invalid
`.changeset/*.md` files

fixes #14023